Create Class/Course implementation, still working on adding in FK

+++ src/main/java/com/revature/project2/controllers/CourseController.java
@@ -2,7 +2,6 @@ package com.revature.project2.controllers;
 
 import com.revature.project2.entities.Course;
 import com.revature.project2.services.CourseServiceImpl;
-import com.revature.project2.services.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,7 +19,7 @@ public class CourseController {
     }
 
     @GetMapping("/courses/{id}")
-    public Course getCourseById(@PathVariable("id") long id){
+    public Course getCourseById(@PathVariable("id") int id){
         return service.getCourseById(id);
     }
 
@@ -30,13 +29,13 @@ public class CourseController {
     }
 
     @PutMapping("/courses/{id}")
-    public String updateCourse(@PathVariable("id") long id, @RequestBody Course updateCourse){
+    public String updateCourse(@PathVariable("id") int id, @RequestBody Course updateCourse){
         service.updateCourse(id, updateCourse);
         return "record updated successfully";
     }
 
     @DeleteMapping("/courses/{id}")
-    public String deleteCourse(@PathVariable("id") long id){
+    public String deleteCourse(@PathVariable("id") int id){
         service.deleteCourse(id);
         return "record deleted successfully";
     }
